+++ src/main/java/com/muhuang/salecrawler/SaleCrawlerApplication.java
@@ -2,7 +2,9 @@ package com.muhuang.salecrawler;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
+@EnableScheduling
 @SpringBootApplication
 public class SaleCrawlerApplication {
