+++ src/main/java/com/muhuang/salecrawler/schedule/ScheduleStatus.java
@@ -1,5 +1,5 @@
 package com.muhuang.salecrawler.schedule;
 
 public enum ScheduleStatus {
-    RUNNING, READY
+    RUNNING, PENDING, FAILED
 }
